Have defaultEmacsDir and defaultServerFile always return something.
It's useful for these to return something even when the directory or
file was not found, because their values are displayed by the -help text
and lets the user know where the command looked for them.

Also extend the tests to cover the case where the server file is in the
config directory.